backend/requirements.txt
# 后端依赖
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
claude-code-sdk>=0.0.25
python-multipart>=0.0.9
pydantic>=2.7.0
pyyaml>=6.0.1
aiofiles>=23.2.1
aiosqlite>=0.19.0
